requirements.txt
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
plotly==5.22.0
selenium==4.30.0
mysql-connector-python==9.2.0
python-dotenv==1.0.1
webdriver-manager==4.0.2
